[FC-6] Built out an error layout with a textview that has a message.
This view will be shown when we catch an error.

Added a circular progress indicator that displays when the data is being loaded.

The viewmodel has a MutableSharedflow called _requestStatusMessage which will be modified if there is an error that has been caught. When there is a change, a snackbar will be fired.

famCardList StateFlow will hold the result of all the cards. We add a stateIn function to it with appropriate parameters that helps us handle configuration changes and background observations in a more effecient way.