[FC-9] Built out the logic for the nested recyclerview adapter.
Added Coil library to load images.

We now have a CardGroupAdapter.kt that acts as the main container for all the child recyclerviews under CardAdapter.kt.

Basic bind functions with CTA actions have been defined. Lambda functions are passed up above from the CardAdapter.kt to handle onClick events of cards.

An extension function was written to animate the slide animation when the big card is long pressed.

Both dismiss and remind buttons removes the card from the list. Have to configure the shared preference class and add the dismissed card id to a list and then filter them out when a new list is being submitted to the adapter on refresh/new app launch.